conda refactor - kamangir/bolt#420

+++ .abcli/conda.sh
@@ -4,7 +4,7 @@ function vancouver_watching_conda() {
     local task=$(abcli_unpack_keyword $1 help)
 
     if [ "$task" == "help" ]; then
-        abcli_show_usage "vanwatch conda create_env [validate]" \
+        abcli_show_usage "vanwatch conda create_env [validate,~recreate]" \
             "create conda environment."
         abcli_show_usage "vanwatch conda validate" \
             "validate conda environment."
@@ -13,9 +13,17 @@ function vancouver_watching_conda() {
 
     if [ "$task" == "create_env" ]; then
         local options=$2
+        local do_recreate=$(abcli_option_int "$options" recreate 1)
         local do_validate=$(abcli_option_int "$options" validate 0)
 
-        abcli_conda create_env name=Vancouver-Watching
+        local environment_name=Vancouver-Watching
+
+        if [[ "$do_recreate" == 0 ]] && [[ $(abcli_conda exists $environment_name) == 1 ]]; then
+            abcli_eval - conda activate $environment_name
+            return
+        fi
+
+        abcli_conda create_env name=$environment_name
 
         pip3 install pymysql==0.10.1
         pip3 install geojson
@@ -37,7 +45,6 @@ function vancouver_watching_conda() {
     fi
 
     if [ "$task" == validate ]; then
-        echo "wip"
         return
     fi
